requirements.txt
streamlit
scikit-learn
openai
openai
python-dotenv
python-dotenv
pywin32
pandas
python-dateutil
plotly
